Move classes to individual files in new surface directory
