+++ resqpy/surface/__init__.py
@@ -1,0 +1,8 @@
+__all__ = ['base_surface', 'combined_surface', 'mesh', 'triangulated_patch', 'pointset', 'surface']
+
+from .base_surface import BaseSurface
+from .combined_surface import CombinedSurface
+from .mesh import Mesh
+from .triangulated_patch import TriangulatedPatch
+from .pointset import PointSet
+from .surface import Surface
