+++ tests/test_data/Charisma_points.txt
@@ -1,0 +1,15 @@
+INLINE :  25701 XLINE :  23693    420691.19624   6292314.22044      2799.05591
+INLINE :  25701 XLINE :  23694    420680.15765   6292308.35532      2798.08496
+INLINE :  25701 XLINE :  23695    420669.11907   6292302.49019      2798.56592
+INLINE :  25701 XLINE :  23696    420658.08049   6292296.62507      2799.31055
+INLINE :  25701 XLINE :  23697    420647.04190   6292290.75995      2799.00537
+INLINE :  25701 XLINE :  23698    420636.00332   6292284.89483      2798.11792
+INLINE :  25701 XLINE :  23699    420624.96473   6292279.02970      2797.75220
+INLINE :  25701 XLINE :  23700    420613.92615   6292273.16458      2797.47607
+INLINE :  25702 XLINE :  22900    429438.92809   6296976.30174      2895.28198
+INLINE :  25702 XLINE :  22901    429427.88950   6296970.43661      2895.15771
+INLINE :  25702 XLINE :  22902    429416.85092   6296964.57149      2895.23901
+INLINE :  25702 XLINE :  22903    429405.81234   6296958.70637      2895.29688
+INLINE :  25702 XLINE :  22904    429394.77375   6296952.84124      2895.31128
+INLINE :  25702 XLINE :  22905    429383.73517   6296946.97612      2895.40430
+INLINE :  25702 XLINE :  22906    429372.69659   6296941.11100      2895.52881
