+++ tests/test_data/IRAP_points.txt
@@ -1,0 +1,9 @@
+429450.658333 6296954.224574 2403.837646
+429444.793211 6296965.263155 2403.449707
+429438.928088 6296976.301737 2403.005615
+429433.062966 6296987.340319 2402.510010
+429427.197843 6296998.378901 2401.972412
+429421.332720 6297009.417483 2401.397949
+429415.467598 6297020.456065 2400.794678
+429409.602475 6297031.494647 2400.166992
+429403.737353 6297042.533228 2399.487793
