+++ tests/test_surface.py
@@ -1,6 +1,6 @@
 import numpy as np
 from numpy.testing import assert_array_almost_equal
-
+import os
 import resqpy.grid
 import resqpy.grid_surface as rqgs
 import resqpy.lines as rql
@@ -11,7 +11,6 @@ import resqpy.surface
 
 
 def test_surface(tmp_model):
-
     # Set up a Surface
     title = 'Mountbatten'
     model = tmp_model
@@ -69,7 +68,6 @@ def test_faces_for_surface(tmp_model):
 
 
 def test_delaunay_triangulation(example_model_and_crs):
-
     model, crs = example_model_and_crs
 
     # number of random points to use
@@ -78,7 +76,7 @@ def test_delaunay_triangulation(example_model_and_crs):
     # create a set of random points
     x = np.random.random(n) * 1000.0
     y = np.random.random(n) * 1000.0
-    z = np.random.random(n)  # note: triangulation does not use z values
+    z = np.random.random(n)  #  note: triangulation does not use z values
     p = np.stack((x, y, z), axis = -1)
 
     # make a PointSet object
@@ -113,10 +111,9 @@ def test_delaunay_triangulation(example_model_and_crs):
 
 
 def test_regular_mesh(example_model_and_crs):
-
     model, crs = example_model_and_crs
 
-    # number of points in mesh, origin spacing
+    #  number of points in mesh, origin spacing
     ni = 7
     nj = 5
     origin = (409000.0, 1605000.0, 0.0)
@@ -173,3 +170,172 @@ def test_regular_mesh(example_model_and_crs):
     assert_array_almost_equal(np.min(p, axis = 0), np.min(peristent_mesh.full_array_ref().reshape(-1, 3), axis = 0))
     assert_array_almost_equal(np.max(p, axis = 0), np.max(peristent_mesh.full_array_ref().reshape(-1, 3), axis = 0))
     assert len(surf.distinct_edges()) == 6 * (ni - 1) * (nj - 1) + (ni - 1) + (nj - 1)
+
+
+def test_points(example_model_and_crs):
+    model, crs = example_model_and_crs
+
+    # Make some random points
+
+    # create some random x,y,z values
+    x = (np.random.random(50) * 10000)
+    y = (np.random.random(50) * 10000)
+    z = (np.random.random(50) * 500 + 2000)
+
+    # make a pointset representation
+    points = resqpy.surface.PointSet(model,
+                                     crs_uuid = crs.uuid,
+                                     points_array = np.array([x, y, z]).T,
+                                     title = 'random points',
+                                     originator = 'Emma',
+                                     extra_metadata = {'testing mode': 'automated'})
+    assert points is not None
+    points.write_hdf5()
+    points.create_xml()
+    points_uuid = points.uuid
+
+    # fully write model to disc
+    model.store_epc()
+    epc = model.epc_file
+
+    # re-open model and check the points object is there
+    model = rq.Model(epc)
+    assert bu.matching_uuids(model.uuid(obj_type = 'PointSetRepresentation', title = 'random points'), points_uuid)
+
+    # establish a resqpy Pointset from the object in the RESQML dataset
+    saved_points = resqpy.surface.PointSet(model, uuid = points_uuid)
+
+    # check a fully expanded version of the points
+    assert_array_almost_equal(saved_points.full_array_ref(), points.full_array_ref())
+
+
+def test_charisma(example_model_and_crs, test_data_path, tmp_path):
+    # Set up a PointSet and save to resqml file
+    model, crs = example_model_and_crs
+
+    charisma_file = test_data_path / "Charisma_points.txt"
+    points = resqpy.surface.PointSet(parent_model = model, charisma_file = str(charisma_file), crs_uuid = crs.uuid)
+    points.write_hdf5()
+    points.create_xml()
+    model.store_epc()
+
+    # Test reload from resqml
+    model = rq.Model(epc_file = model.epc_file)
+    reload = resqpy.surface.PointSet(parent_model = model, uuid = points.uuid)
+
+    assert reload.title == str(charisma_file)
+
+    coords = reload.full_array_ref()
+    assert_array_almost_equal(coords[0], np.array([420691.19624, 6292314.22044, 2799.05591]))
+
+    assert coords.shape == (15, 3), f'Expected shape (15,3), not {coords.shape}'
+
+    # Test write back to file
+    out_file = str(tmp_path / "Charisma_points_out.txt")
+    reload.convert_to_charisma(out_file)
+
+    assert os.path.exists(out_file)
+    with open(out_file, 'r') as f:
+        line = f.readline()
+
+    assert line == 'INLINE :\t1 XLINE :\t1\t420691.19624\t6292314.22044\t2799.05591\n', 'Output Charisma file does not look as expected'
+
+
+def test_irap(example_model_and_crs, test_data_path, tmp_path):
+    # Set up a PointSet and save to resqml file
+    model, crs = example_model_and_crs
+    irap_file = test_data_path / "IRAP_points.txt"
+    points = resqpy.surface.PointSet(parent_model = model, irap_file = str(irap_file), crs_uuid = crs.uuid)
+    points.write_hdf5()
+    points.create_xml()
+    model.store_epc()
+
+    # Test reload from resqml
+    model = rq.Model(epc_file = model.epc_file)
+    reload = resqpy.surface.PointSet(parent_model = model, uuid = points.uuid)
+
+    assert reload.title == str(irap_file)
+
+    coords = reload.full_array_ref()
+    assert_array_almost_equal(coords[0], np.array([429450.658333, 6296954.224574, 2403.837646]))
+
+    assert coords.shape == (9, 3), f'Expected shape (9,3), not {coords.shape}'
+
+    # Test write back to file
+    out_file = str(tmp_path / "IRAP_points_out.txt")
+    reload.convert_to_irap(out_file)
+
+    assert os.path.exists(out_file)
+    with open(out_file, 'r') as f:
+        line = f.readline()
+
+    assert line == '429450.658333 6296954.224574 2403.837646\n', 'Output IRAP file does not look as expected'
+
+
+def test_from_polyline(example_model_and_crs):
+    # Set up a PolyLine and save to resqml file
+    model, crs = example_model_and_crs
+
+    coords = np.array([[1, 2, 3, 4, 5], [6, 7, 8, 9, 10], [12, 13, 14, 15, 16]]).T
+    lines = resqpy.lines.Polyline(parent_model = model,
+                                  set_coord = coords,
+                                  title = 'Polylines',
+                                  set_crs = crs.uuid,
+                                  set_bool = False)
+    lines.write_hdf5()
+    lines.create_xml()
+    model.store_epc()
+
+    # Reload the model, and generate pointset using the polyline
+    model = rq.Model(epc_file = model.epc_file)
+    reload_lines = resqpy.lines.Polyline(parent_model = model, uuid = lines.uuid)
+
+    points = resqpy.surface.PointSet(parent_model = model, polyline = reload_lines, crs_uuid = crs.uuid)
+    points.write_hdf5()
+    points.create_xml()
+    model.store_epc()
+
+    # Reload the model, and ensure the coordinates are as expected
+    model = rq.Model(epc_file = model.epc_file)
+    reload = resqpy.surface.PointSet(parent_model = model, uuid = points.uuid)
+    assert_array_almost_equal(reload.full_array_ref(), coords)
+
+
+def test_from_polylineset(example_model_and_crs):
+    # Set up a PolyLine and save to resqml file
+    model, crs = example_model_and_crs
+
+    coords1 = np.array([[1, 2, 3, 4, 5], [6, 7, 8, 9, 10], [12, 13, 14, 15, 16]]).T
+    coords2 = np.array([[11, 12, 13, 14, 15], [16, 17, 18, 19, 110], [112, 113, 114, 115, 116]]).T
+
+    line1 = resqpy.lines.Polyline(parent_model = model,
+                                  title = 'Polyline1',
+                                  set_coord = coords1,
+                                  set_crs = crs.uuid,
+                                  set_bool = False)
+
+    line2 = resqpy.lines.Polyline(parent_model = model,
+                                  title = 'Polyline2',
+                                  set_coord = coords2,
+                                  set_crs = crs.uuid,
+                                  set_bool = False)
+
+    lines = resqpy.lines.PolylineSet(parent_model = model, title = 'Polylines', polylines = [line1, line2])
+    lines.write_hdf5()
+    lines.create_xml()
+    model.store_epc()
+
+    # Reload the model, and generate pointset using the polylineset
+    model = rq.Model(epc_file = model.epc_file)
+    reload_lines = resqpy.lines.PolylineSet(parent_model = model, uuid = lines.uuid)
+
+    points = resqpy.surface.PointSet(parent_model = model, polyset = reload_lines, crs_uuid = crs.uuid)
+    points.write_hdf5()
+    points.create_xml()
+    model.store_epc()
+
+    # Reload the model, and ensure the coordinates are as expected
+    model = rq.Model(epc_file = model.epc_file)
+    reload = resqpy.surface.PointSet(parent_model = model, uuid = points.uuid)
+
+    assert_array_almost_equal(reload.full_array_ref(), np.concatenate((coords1, coords2), axis = 0))
